requirements.txt
asgiref==3.9.1
Django==5.2.5
django-debug-toolbar==6.0.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
